Added unit test for SyncCollectionHandler. Moved CollectionTestBase to a separate file.
